Split out behavior to keep per-category code isolated
