+++ tredis/sets.py
@@ -1,0 +1,8 @@
+from tornado import concurrent
+
+from tredis import base
+from tredis import exceptions
+
+
+class Sets(base.Category):
+    pass
